Refactored database and track list classes
Changed:
- Track data is now cached in database rather than loaded from spotify
servers each time app is opened. Greatly reduces loading time and allows
for offline use.
- Syncing with spotify can now be done by tapping a button on the main
screen
- Sync is performed automatically when app is opened for the very first
time
- SourceTrackData heavily refactored, is now a singleton and has new
methods for caching/loading track data
- DownloadTrackDataTask renamed to SyncTrackDataTask, also heavily
refactored to perform new syncing functionality
- Boolean change flags have been removed from Tags
- EditTagActivity no longer returns a result, changes are made directly
from the dialog
- TableRatings renamed to TableTracks, additional columns added
- Other minor tweaks to database table structure
- Database version is now stored in sharedprefs
- Some framework started for playlist generation, to be fully added in
next commit

+++ app/src/main/java/com/kwohlford/smartplaylistmanager/db/SourceTrackData.java
@@ -9,69 +9,156 @@ import android.util.Log;
 
 import com.kwohlford.smartplaylistmanager.tracklist.Tag;
 import com.kwohlford.smartplaylistmanager.tracklist.TrackData;
+import com.kwohlford.smartplaylistmanager.tracklist.TrackListing;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
 
 /**
- * Convenience methods for retrieving data from the trackdata database.
+ * Convenience methods for retrieving data from the trackdata database. Must be opened before
+ * any methods will return data.
  */
 public class SourceTrackData extends DBDataSource {
 
+    private static final String TAG = "Database";
+
+    // Singleton instance of datasource
+    private static SourceTrackData instance;
+
     // Basic database info
     public static final String DB_NAME = "trackdata.db";
-    public static final int DATABASE_VERSION = 5;
 
-    // Cached lists of retrieved tags to reduce database calls
-    private ArrayList<Tag> cachedGenreTags;
-    private ArrayList<Tag> cachedMoodTags;
+    // Whether or not the database has been opened
+    private boolean open;
 
-    // True if tag caches need to be updated, false otherwise
-    private boolean updateCachedTags;
+    // Cached track list (reduces db calls)
+    public TrackListing tracks;
 
-    public SourceTrackData(Context context) {
-        super(context,
-                DB_NAME,
-                DATABASE_VERSION,
-                new TableTags(), new TableTrackTags(), new TableRatings());
-        updateCachedTags = true;
-    }
-
-    @Override
-    public void open() throws SQLException {
-        super.open();
-        updateCachedTags = true;
+    /**
+     * Private constructor to prevent outside instantiation.
+     */
+    private SourceTrackData() {
+        open = false;
     }
 
     /**
+     * @return Singleton instance of the data source
+     */
+    public static SourceTrackData getInstance() {
+        if(instance == null) {
+            instance = new SourceTrackData();
+        }
+        return instance;
+    }
+
+    /**
+     * Open the database. Must be called before any other methods will return data.
+     * @param context Context
+     * @throws SQLException If the database cannot be accessed
+     */
+    public void open(Context context, int version) throws SQLException {
+        super.open(context,
+                DB_NAME,
+                version,
+                new TableTags(), new TableTrackTags(), new TableTracks());
+        open = true;
+        tracks = loadAllTracks();
+    }
+
+    /**
+     * Loads complete list of tracks from the db.
+     * @return List of all tracks and their data
+     */
+    private TrackListing loadAllTracks() {
+        Cursor results = database.query(
+                TableTracks.NAME,
+                null, null, null, null, null, null
+        );
+        results.moveToFirst();
+
+        HashMap<String,TrackData> tracks = new HashMap<>();
+        while(!results.isAfterLast()) {
+            String uri = results.getString(0);
+            TrackData track = new TrackData(
+                    uri,
+                    results.getString(1),
+                    results.getString(2),
+                    results.getString(3),
+                    results.getString(4),
+                    results.getString(5)
+            );
+            track.rating = results.getFloat(6);
+            track.setTags(Tag.TagType.GENRE, getTrackTags(uri, Tag.TagType.GENRE));
+            track.setTags(Tag.TagType.MOOD, getTrackTags(uri, Tag.TagType.MOOD));
+            tracks.put(uri, track);
+            results.moveToNext();
+        }
+        results.close();
+
+        TrackListing trackList = new TrackListing(tracks);
+        trackList.setTags(Tag.TagType.GENRE, getTagsByType(Tag.TagType.GENRE));
+        trackList.setTags(Tag.TagType.MOOD, getTagsByType(Tag.TagType.MOOD));
+
+        return trackList;
+    }
+
+    /**
+     * Adds a new track to the database.
+     * @param track Track to add
+     */
+    public void addTrack(TrackData track) {
+        if(!open) return;
+
+        Log.d(TAG, "Adding track " + track.songTitle);
+        ContentValues values = new ContentValues();
+        values.put(TableTracks.COL_ID, track.uri);
+        values.put(TableTracks.COL_TITLE, track.songTitle);
+        values.put(TableTracks.COL_ARTIST, track.artist);
+        values.put(TableTracks.COL_ALBUM, track.albumName);
+        values.put(TableTracks.COL_PREVIEW, track.previewUrl);
+        values.put(TableTracks.COL_ALBUMART, track.albumArtUrl);
+        values.put(TableTracks.COL_RATING, track.rating);
+        database.insert(
+                TableTracks.NAME,
+                null,
+                values
+        );
+
+        tracks.addTrack(track);
+    }
+
+    /**
+     * Removes a track from the database.
+     * @param track Track to delete
+     */
+    public void deleteTrack(TrackData track) {
+        if(!open) return;
+
+        // Delete all associations with the track
+        database.delete(
+                TableTrackTags.NAME,
+                TableTrackTags.COL_TRACKID + " = ?",
+                new String[]{track.uri}
+        );
+
+        // Delete the track itself
+        Log.d(TAG, "Deleting track " + track.songTitle);
+        database.delete(
+                TableTracks.NAME,
+                TableTracks.COL_ID + " = ?",
+                new String[]{track.uri}
+        );
+        tracks.removeTrack(track);
+    }
+
+    /**
+     * Queries the database to retrieve a list of all available tags in a given category.
      * @param type Category of tags to retrieve
      * @return List of all available tags in the category
      */
-    public ArrayList<Tag> getTagsByType(Tag.TagType type) {
-        if(updateCachedTags) {
-            cachedGenreTags = updateCachedTags(Tag.TagType.GENRE);
-            cachedMoodTags = updateCachedTags(Tag.TagType.MOOD);
-            updateCachedTags = false;
-        }
+    private ArrayList<Tag> getTagsByType(Tag.TagType type) {
+        if(!open) return new ArrayList<>();
 
-        switch(type) {
-            case GENRE:
-                return cachedGenreTags;
-            case MOOD:
-                return cachedMoodTags;
-            default:
-                return new ArrayList<>();
-        }
-    }
-
-    /**
-     * Queries the database to update the cached tag lists.
-     * @param type Category of tags to query for
-     * @return Updated list of tags
-     */
-    private ArrayList<Tag> updateCachedTags(Tag.TagType type) {
         Cursor results = database.query(
                 TableTags.NAME,
                 new String[] { TableTags.COL_NAME },
@@ -94,14 +181,18 @@ public class SourceTrackData extends DBDataSource {
      * @param tag Tag to add
      */
     public void addTag(Tag tag) {
-        updateCachedTags = true;
+        if(!open) return;
+
+        Log.d(TAG, "Adding tag " + tag.name);
         ContentValues values = new ContentValues();
         values.put(TableTags.COL_NAME, tag.name);
         values.put(TableTags.COL_CATEGORY, tag.type.id);
         database.insert(
                 TableTags.NAME,
                 null,
-                values);
+                values
+        );
+        tracks.addTag(tag);
     }
 
     /**
@@ -109,41 +200,69 @@ public class SourceTrackData extends DBDataSource {
      * @param tag Tag to delete
      */
     public void deleteTag(Tag tag) {
-        updateCachedTags = true;
+        if(!open) return;
+
+        // Remove from cached track list
+        ArrayList<String> tracksWithTag = getTrackUrisWithTag(tag);
+        for(String s : tracksWithTag) {
+            tracks.getTrackForUri(s).getTags(tag.type).remove(tag);
+        }
+        tracks.deleteTag(tag);
+
         // Delete all associations with the tag
         int tagId = lookupTagKey(tag);
         database.delete(
                 TableTrackTags.NAME,
                 TableTrackTags.COL_TAGID + " = ?",
-                new String[] { String.valueOf(tagId) }
+                new String[]{String.valueOf(tagId)}
         );
 
         // Delete the tag itself
+        Log.d(TAG, "Deleting tag " + tag.name);
         database.delete(
                 TableTags.NAME,
                 TableTags.COL_ID + " = ?",
-                new String[] { String.valueOf(tagId) }
+                new String[]{String.valueOf(tagId)}
         );
     }
 
     /**
      * Renames a tag in the database.
      * @param prev Tag to be updated
-     * @param edited New tag to replace it with
+     * @param newName New name for the tag
      */
-    public void editTag(Tag prev, Tag edited) {
-        updateCachedTags = true;
+    public void editTag(Tag prev, String newName) {
+        if(!open) return;
+
+        // Update name in cached track list
+        ArrayList<String> tracksWithTag = getTrackUrisWithTag(prev);
+        for(String s : tracksWithTag) {
+            ArrayList<Tag> trackTags = tracks.getTrackForUri(s).getTags(prev.type);
+            trackTags.get(trackTags.indexOf(prev)).name = newName;
+        }
+        ArrayList<Tag> allTags = tracks.getTags(prev.type);
+        allTags.get(allTags.indexOf(prev)).name = newName;
+
+        // Update in database
+        Log.d("Edit Tags", "Renaming tag " + prev.name + " to " + newName);
         ContentValues values = new ContentValues();
-        values.put(TableTags.COL_NAME, edited.name);
+        values.put(TableTags.COL_NAME, newName);
         database.update(
                 TableTags.NAME,
                 values,
                 TableTags.COL_NAME + " = ? AND " + TableTags.COL_CATEGORY + " = ?",
-                new String[] { prev.name, String.valueOf(prev.type.id) }
+                new String[]{prev.name, String.valueOf(prev.type.id)}
         );
     }
 
+    /**
+     * Queries the database to retrieve a list of tracks which are associated with a tag.
+     * @param tag Tag to search for
+     * @return List of tracks associated with the tag
+     */
     public ArrayList<String> getTrackUrisWithTag(Tag tag) {
+        if(!open) return new ArrayList<>();
+
         String query =
                 "SELECT track_tags.track_id " +
                         "FROM track_tags " +
@@ -151,7 +270,7 @@ public class SourceTrackData extends DBDataSource {
                         "WHERE tags.category = ? AND tags.name = ?";
         Cursor results = database.rawQuery(
                 query,
-                new String[] { String.valueOf(tag.type.id), tag.name }
+                new String[]{String.valueOf(tag.type.id), tag.name}
         );
         results.moveToFirst();
 
@@ -165,39 +284,22 @@ public class SourceTrackData extends DBDataSource {
     }
 
     /**
-     * Queries the database to get a track's rating.
-     * @param trackUri Uri of track to search for
-     * @return User-defined rating of the track (or 0 if none has been set)
-     */
-    public float getRating(String trackUri) {
-        Cursor results = database.query(
-                TableRatings.NAME,
-                new String[]{ TableRatings.COL_RATING },
-                TableRatings.COL_ID + " = ?",
-                new String[]{ trackUri },
-                null, null, null
-        );
-        results.moveToFirst();
-        float rating = results.isAfterLast() ? 0 : results.getFloat(0);
-        results.close();
-        return rating;
-    }
-
-    /**
-     * Inserts or overwrites a track's rating in the database.
+     * Updates a track's rating in the database.
      * @param trackUri Track to set rating for
      * @param rating User rating out of 5
      */
     public void setRating(String trackUri, float rating) {
+        if(!open) return;
+
         ContentValues values = new ContentValues();
-        values.put(TableRatings.COL_ID, trackUri);
-        values.put(TableRatings.COL_RATING, rating);
-        Log.d("Database", "Setting rating " + rating + " on track " + trackUri);
-        database.insertWithOnConflict(
-                TableRatings.NAME,
-                null,
+        values.put(TableTracks.COL_RATING, rating);
+        Log.d(TAG, "Setting rating " + rating + " on track " + trackUri);
+
+        database.update(
+                TableTracks.NAME,
                 values,
-                SQLiteDatabase.CONFLICT_REPLACE
+                TableTracks.COL_ID + " = ?",
+                new String[]{trackUri}
         );
     }
 
@@ -208,6 +310,8 @@ public class SourceTrackData extends DBDataSource {
      * @return List of tags associated with the track
      */
     public ArrayList<Tag> getTrackTags(String trackUri, Tag.TagType type) {
+        if(!open) return new ArrayList<>();
+
         String query =
                 "SELECT tags.name " +
                         "FROM track_tags " +
@@ -236,15 +340,17 @@ public class SourceTrackData extends DBDataSource {
      * @param deletedEntries Tags which have been deleted since the last update
      */
     public void setTrackTags(String trackUri, ArrayList<Tag> addedEntries, ArrayList<Tag> deletedEntries) {
+        if(!open) return;
+
         for(Tag t : addedEntries) {
             int tagKey = lookupTagKey(t);
-            Log.d("Database", "Adding tag " + t.name + " to track " + trackUri);
+            Log.d(TAG, "Adding tag " + t.name + " to track " + trackUri);
             addTrackTag(trackUri, tagKey);
         }
 
         for(Tag t : deletedEntries) {
             int tagKey = lookupTagKey(t);
-            Log.d("Database", "Deleting tag " + t.name + " from track " + trackUri);
+            Log.d(TAG, "Deleting tag " + t.name + " from track " + trackUri);
             deleteTrackTag(trackUri, tagKey);
         }
     }
@@ -255,6 +361,8 @@ public class SourceTrackData extends DBDataSource {
      * @param tagKey PK of tag entry in 'tags' table
      */
     private void addTrackTag(String trackUri, int tagKey) {
+        if(!open) return;
+
         ContentValues values = new ContentValues();
         values.put(TableTrackTags.COL_TRACKID, trackUri);
         values.put(TableTrackTags.COL_TAGID, tagKey);
@@ -271,6 +379,8 @@ public class SourceTrackData extends DBDataSource {
      * @param tagKey PK of tag entry in 'tags' table
      */
     private void deleteTrackTag(String trackUri, int tagKey) {
+        if(!open) return;
+
         database.delete(
                 TableTrackTags.NAME,
                 TableTrackTags.COL_TAGID + " = ? AND " + TableTrackTags.COL_TRACKID + " = ?",
